+++ pjsonlib/src/pjson_schema_format.cpp
@@ -2,6 +2,7 @@
 // Copyright 2025 ByteDance Ltd. and/or its affiliates. All rights reserved.
 // Licensed under the Apache License, Version 2.0 (the "License");
 //
+#include "pjson_schema_regex.h"
 #include "pjson_schema_util.h"
 
 #include <cstddef>
@@ -196,6 +197,8 @@ namespace ByteDance {
                 return validIPv6(value);
             if (format == "uuid")
                 return validUuid(value);
+            if (format == "regex")
+                return validEcmaRegex(value);
             known = false;
             return true;
         }